Set has almost been upgraded.
