+++ lib/falcor/operations/index.js
@@ -1,7 +1,7 @@
 var ModelResponse = require('../ModelResponse');
-var get = require('./get');
-var set = require('./set');
-var invalidate = require('./invalidate');
+var request = require('./request');
+var setInitialArgs = require('./set/buildInitialArgs');
+var setSourceRequest = require('./set/modelSourceRequest');
 
 module.exports = function modelOperation(name) {
     return function() {
@@ -18,7 +18,7 @@ module.exports = function modelOperation(name) {
         if (name === 'get') {
             modelResponder = get(model, args, selector);
         } else if (name === 'set') {
-            modelResponder = set(model, args, selector);
+            modelResponder = request(model, args, selector, setInitialArgs, setSourceRequest);
         } else if (name === 'invalidate') {
             modelResponder = invalidate(model, args, selector);
         }
